+++ R/TMB1964r.R
@@ -77,7 +77,7 @@
 #'     variables = c("T1","T2","T3","T4","T5","T6","T7"),
 #'     adjustments = list(purpose="difference", decorrelation="CM"),
 #'     plotStyle = "line", 
-#'     errorbarParams = list(size=0.05, position = position_dodge(.5) ),
+#'     errorbarParams = list(size=0.15, position = position_dodge(.5) ),
 #'     pointParams = list(size=2.5, position = position_dodge(.5)),
 #'     lineParams = list(size=0.25)
 #' ) +