+++ R/functionsPlotting_base.R
@@ -375,6 +375,9 @@ superbPlot.pointjitter <- function(
     runDebug("pointjitter", "Entering superbPlot.pointjitter", 
         c("xfactor2", "groupingfactor2", "addfactors2","pointParams2","jitterParams2","errorbarParams2"), list(xfactor, groupingfactor, addfactors, pointParams, jitterParams, errorbarParams))
 
+    # rename column "DV" to "center"
+    rawdata$center <- rawdata$DV
+
     # depending on the scale of the x-axis.
     if (!xAsFactor) {
         summarydata[[xfactor]] = as.numeric(summarydata[[xfactor]])
@@ -385,14 +388,14 @@ superbPlot.pointjitter <- function(
     if (is.null(groupingfactor)) {
         do_jitters = do.call(geom_jitter, modifyList(
                         list(data = rawdata, alpha = 0.2, width = 0.2, height = 0.0,
-                             mapping = aes_string(y = "DV" ) ),
+                             mapping = aes_string(y = "center" ) ),
                         jitterParams
                     ) )
     } else {
         do_jitters = do.call(geom_point, modifyList(
                         list(data = rawdata , alpha = 0.2,
                             position = position_jitterdodge(jitter.width=0.1 , dodge.width=0.5 ),
-                            mapping = aes_string(y = "DV", color = groupingfactor  ) ),
+                            mapping = aes_string(y = "center", color = groupingfactor  ) ),
                         jitterParams
                     ) )
     }
@@ -492,16 +495,19 @@ superbPlot.pointjitterviolin <- function(
     runDebug("pointjitterviolin", "Entering superbPlot.pointjitterviolin", 
         c("xfactor2", "groupingfactor2", "addfactors2","pointParams2","jitterParams2","violinParams2","errorbarParams2"), list(xfactor, groupingfactor, addfactors, pointParams, jitterParams, violinParams, errorbarParams))
 
+    # rename column "DV" as "center"
+    rawdata$center <- rawdata$DV
+
     # determining the type of jitter based on the presence or not of a groupingfac
     if (is.null(groupingfactor)) {
         do_jitters = do.call(geom_jitter, modifyList(
                         list(data = rawdata, alpha = 0.2, width = 0.2, height = 0.0,
-                             mapping = aes_string( y = "DV" ) ),
+                             mapping = aes_string( y = "center" ) ),
                         jitterParams
                     ) )
         do_violins = do.call( geom_violin, modifyList(
                         list(data     = rawdata,
-                             mapping  = aes_string( y = "DV" ), 
+                             mapping  = aes_string( y = "center" ), 
                              scale    = "area", trim = FALSE, alpha = 0.25),
                         violinParams
                     ) )
@@ -509,13 +515,13 @@ superbPlot.pointjitterviolin <- function(
         do_jitters = do.call(geom_point, modifyList(
                         list(data = rawdata , alpha = 0.2,
                             position = position_jitterdodge(jitter.width=0.1 , dodge.width=.75 ),
-                            mapping = aes_string(y = "DV", group = groupingfactor  ) ),
+                            mapping = aes_string(y = "center", group = groupingfactor  ) ),
                         jitterParams
                     ) )
         do_violins = do.call( geom_violin, modifyList(
                         list(data    = rawdata, 
                              position= position_dodge(.75), #"dodge",
-                             mapping = aes_string( y = "DV", fill = groupingfactor), 
+                             mapping = aes_string( y = "center", fill = groupingfactor), 
                              scale   = "area", trim = FALSE, alpha = 0.25),
                         violinParams
                     ) )
@@ -525,11 +531,11 @@ superbPlot.pointjitterviolin <- function(
     plot <- ggplot(data    = summarydata, 
                    mapping = aes_string(x = xfactor, colour = groupingfactor )
         ) +
-        # violin in the back
+        # violins in the back
         do_violins +
-        # jitter second
+        # jitters second
         do_jitters +
-        # and finally the point and the error bars
+        # and finally the points and the error bars
         do.call( geom_point, modifyList(
             list(mapping = aes_string(group = groupingfactor, y = "center"), 
                 size = 3, position = position_dodge(.75) ),
@@ -624,6 +630,9 @@ superbPlot.pointindividualline <- function(
     runDebug("pointindividualline", "Entering superbPlot.pointindividualline", 
         c("xfactor2", "groupingfactor2", "addfactors2","pointParams2","lineParams2","errorbarParams2"), list(xfactor, groupingfactor, addfactors, pointParams, lineParams, errorbarParams))
 
+    # rename column "DV" as "center"
+    rawdata$center <- rawdata$DV
+
     # let's do the plot!
     plot <- ggplot(
         data = summarydata, 
@@ -635,13 +644,13 @@ superbPlot.pointindividualline <- function(
     do.call(geom_line, modifyList(
         list(data = rawdata,
             size=0.2, alpha = 0.25,
-            mapping = aes_string( y = "DV", group = "id" ) ),
+            mapping = aes_string( y = "center", group = "id" ) ),
         lineParams
     )) +
     # the individual points 
     do.call(geom_point, modifyList(
         list(data = rawdata, alpha = 0.25,
-            mapping = aes_string(y = "DV", group = "id") ),
+            mapping = aes_string(y = "center", group = "id") ),
         pointParams
     )) + 
     # the points